Move factory methods to a separate folder
