+++ packages/hooks/src/hooks/usePairPrice.ts
@@ -1,4 +1,4 @@
-import { getPairPrice } from 'core'
+import { actions } from 'core'
 import { BlockNumber, BlockTag } from 'starknet'
 
 import { Pair } from '../types'
@@ -12,7 +12,7 @@ export const usePairPrice = (pair?: Pair, blockNumber: BlockNumber = BlockTag.la
     queryKey: ['pairPrice', pair?.address, pair?.reversed],
     queryFn: async () => {
       if (!pair) return
-      return getPairPrice(factory.config.provider, pair, blockNumber)
+      return actions.getPairPrice(factory.config, pair, blockNumber)
     },
     enabled: !!pair,
   })
